This adds support for authenticated docker_pull and docker_push.
This changes the mechanics of pull to use a prebuilt docker_puller PAR.  This is so that the pull can properly take place during WORKSPACE instantiation and not during the build.  This is so that the docker_puller can access the Docker credentials configured and invoke credential helpers on the host filesystem.

The new docker_push path utilizes a prebuilt docker_pusher in a similar manner to the revised approach above.