This implements fast docker_{push,pull}
This commit pulls in the latest release of `google/containerregistry`, which includes a new fast-path for puller/pusher to use, which is consistent with our new intermediate format.

This commit adapts `docker_pull` to consume the new puller and stitch together the image's constituent elements using `docker_import`.

This commit also changes `docker_push` to consume the new options on the new pusher to publish images without linking the full `docker save` tarball, which can be quite expensive for large images.