Rework the incremental loading logic, and expand e2e testing.
I uncovered an issue as I expanded the testing of `bazel run` of the build targets.  The `docker import` we were performing previously only makes the imported diff id available as the bottom-most base layer.  It does this by looking at the diff-id within the context of its base image "chain".

So this reworks the existence check logic to effective probe the diff-id chain and then synthesize a tarball of all remaining layers, which is imported directly into the Docker daemon.

With these changes, the new tests in ./testing/e2e.sh pass for me locally, where before they failed on the `docker load`