Migrate away from depending on a pusher.par.
This migrates push code away from using the published pusher.par file in favor of consuming the tool's BUILD target directly.  AIUI based on the discussion [here](https://github.com/bazelbuild/rules_docker/issues/136#issuecomment-329528347) this should make things more portable because the `py_binary` target should not need to be prefixed with a Python executable on Windows.

This also simplifies the process of updating the `containerregistry` dependency (fewer hashes), but incurs a slightly more expensive build for push targets (this is probably ok because it will be cached).