Bring sanity to java_image paths (#156)
* Hoist some boilerplate code into a function.

* Bring sanity to java_image layouts.

This starts to share with `java_image` some of the `file_map` logic which brought sanity to the layout of `py_image` (fixing issues with external dependencies).

Now, the `java_image` dependency layer logic is almost completely shared with the common `dep_layer_impl`, the one difference is that I've made the way runfiles are accessed pluggable because Java handles this differently than other languages.  The `java_image` app layer uses a number of the same utilities as the refactored common logic, but remains separate so that the classpath can be properly handled (as well as Java-specific attributes like `runtime_deps`).

The WAR logic remains independent of all of this.  Because we build an "exploded WAR" layout things must be flattened, so it wasn't as crazy as its siblings.  However, this adds a new TODO to leverage the `file_map` functionality to avoid the potential collisions that could happen today when the basename of two `java_library` dependencies match.

Fixes: https://github.com/bazelbuild/rules_docker/issues/106

* Fix buildifier issues.