Add a py3_image rule
This rule is an analog to `py_image`, only the resulting image runs
Python 3 instead of Python 2.  I'm certainly open to suggestions on
a more elegant way of doing this using attributes on the stock
`py_image` rule, but I'm not sure if there is one.

This also updates a bunch of unrelated stuff as a consequence of running
`update_deps.sh`.

This fixes bazelbuild/rules_docker#229