Add support for print_progress to container_pull
Add minimal prints to indicate puller progress if container_pull
has print_progress=True attribute.  The intention is that the pull
operation would provide some feedback to the user, instead of bazel
getting seemingly stuck for 10 minutes when pulling an image that is
many gigabytes in size.

Pending pull request to upstream puller.par is here:
https://github.com/google/containerregistry/pull/66