A capability to add empty directories to a container image (#310)
* Fix incorrect docstring.

The docstring for "mode" param for "add_empty_file" seems to have been
cargo-culted from "add_file" function. The default mode is hardcoded
0o644, rather than the mode of the source file (there is no source file
in the context of "add_empty_file" in the first place).

* Add "empty_dirs" parameter for container_image.

A capability to add an empty directory with specific mode to a container
image. The specific usecase is to be able to create /tmp directory.

One small side-effect of this change is, that now --empty_file flag for
build_tar.py obeys --mode and --mores flags.

* style: use "dict.get(val, default)" rather than "if val in dict: dict[val]"