Fix runfiles discovery
The launcher scripts were not able to find runfiles when executed
deeper inside a runfile tree generated by another rule.  This
happens when another rule depends on the container image target
and wants to execute the scripts from its own runfiles tree.

Fix discovery to also account for this possibility, using similar
heuristics as FindModuleSpace in the python launcher.