Make non-hermetic gzip/gunzip more portable (#373)
This ought to make the gzip and gunzip actions work on systems where
the gzip binary is in a path that would not normally show up in the
action sandbox.

For example, on NixOS the gzip binary is typically at
/run/current-system/sw/bin/gzip (or possibly elsewhere if a user has
installed it into their own profile).  While these locations are
definitely present in $PATH, Bazel does not normally make them
available in execution sandboxes.

There should be no noticeable difference on any system where these
rules were already working; this change broadens support and makes the
non-hermetic steps explicit.