Add a `null_entrypoint` attribute to app_layer
This allows people to not use the default entrypoint which is geared
towards runnig the container as executable with one or more arguments.

With zero arguments also, it is a problem when the base image already
specifies a Cmd in its config. Because there is no way to override
the config currently.

See https://github.com/google/containerregistry/pull/86