Use the zipped layer to calculate blobsums on imports (#478)
* Use the zipped layer to calculate blobsums on imports

  If a docker save tarball was imported with unzipped layers the blobsum
  would incorrectly be calculated off of the unzipped layer and cause
  digest mismatches when pushing.

* Rehome container_import testdata and fix container/image_test.py

Both testdata/pause_piecemeal and pause/piece_piecemeal_gz are
generating identical blobsum artifacts from separate actions, which
bazel does not allow during the analysis phase.

This is an an issue now because the container_import rule generates consistent
blobsums for either gzipped or unzipped layers. This commit moves the
import testdata to separate packages to workaround this.