Java data runfiles (#530)
Add examples using data dependencies with Java the runfiles library. These demonstrate the problem described in #506.

Fix the java_image rule so that these examples work. This required an additional environment variable being passed when executing the binary, and adding the data dependencies to the files that are part of the image.

This in turn required adding the _jdk attribute, so the files that appear to be data dependencies but come from the JRE may be excluded by marking them as already available.

NOTE:  data files of dependencies are left in the top layer -- to get things working for the time being. The disadvantage is that a change in dependencies may require rebuilding both the dependency layer and the top layer (I think), but I'm not sure of the cost of that. It also means that every time the top layer changes, the data dependencies need to be re-tar'd, etc. Again, likely not a huge problem, and something that seems better to fix in the future by adding appropriate exclusion mechanisms to the rules that build the dep layer.