Allow container_layer's data_path to work on directory output target names.
There is a tiny bug in pkg_tar's strip_prefix handling that prevents it
from stripping the full name of a file. Though this sounds weird at
first, it is pretty useful when creating tarballs from output
directories. A pull request for this has been sent to Bazel:

    https://github.com/bazelbuild/bazel/pull/6980

The same issue applies to container_layer's data_path argument, which is
actually where we noticed this at first. Also apply this fix here.