Make obtainer build user in tests remote execution compatible
image_test.py currently uses the environment variable USER to obtain the
build user. While this matches the BUILD_USER when bazel is invoked
localy, it can be different when the test action is executed in a remote
machine. This commit makes the Bazel stamp variables available to
image_test.py so that it can obtain the BUILD_USER as set by Bazel.