Do not add --all_incompatible_changes to presubmit (#662)
The incompatible changes is a little bit more involved than this, we
allow creating new incompatible flags which we don't expect users to
migrate for right away. The principled place to look for incompatibel
changes is to go to https://github.com/bazelbuild/bazel/issues?q=is%3Aissue+label%3Aincompatible-change+label%3Amigration-ready,
ideally also filtered with the bazel version (e.g. for flags that should
be migrated in bazel 0.22.0, add the label 'migration-0.22'.

On the buildkite, this is the pipeline that tells rule owners which
incompatible flags they have to migrate for in the current bazel
release: https://buildkite.com/bazel/bazel-at-release-plus-incompatible-flags