Expose attributes from lang/image.bzl as struct
Besides exposing the rule itself, it is a common pattern to expose
rule attributes, outputs, toolchains and the implementation function
as a struct.  This makes it easier to wrap the rule in another rule,
without duplicating the attribute definitions.  Apply this pattern
to lang/image.bzl, and remove duplicated definitions from the wrapping
rules in java/image.bzl and nodejs/image.bzl.