Allow env args to docker_image that contain "="
The current implementation of this is designed to split key-value pairs
using an equal sign, but this behavior can interfere with pairs where
the value itself contains an equal sign. It seems that the correct
behavior is actually to split only once (preserving any equal signs
after the first one)  and then make sure that there is
at least one valid pair.