config_stripper no longer stores full file in RAM.
Previously, config_stripper would store the full
TAR file and gzipped TAR file in memory to compute
the SHA256 hash.

After this change, the stripped TAR file is stored
in a temporary location on disk and then renamed
into its final location.  Furthermore, the SHA256
hash of the gzipped TAR file is computed
incrementally as the gzipping progresses.