install_pkgs uses named volumes to work with DIND.
Previously, install_pkgs would mount files into
the target image directly (with '-v src:dst').
This unfortunately doesn't work with
docker-in-docker (DIND) setups where the mount
is done from the perspective of the host which
does not share a filesystem with the docker client.

To work with such setups, install_pkgs now creates
a named volume and copies files into the named
volume.  The named volume containing the install
script and TAR file is then mounted into the
target container.