Use 'portable' mtime as default for tars, like rules_pkg does
Support 'portable' mtime as option for container layer tars, like
rules_pkg does

Originally did this in a backwards incompatible way, but this should be
backwards compatible now.

I should've done it like rules_pkg does from the start:

https://github.com/bazelbuild/rules_pkg/blob/master/pkg/pkg.bzl#L275-L276

The difference here is we don't default to 'portable', since that'd
break backwards compatibility, so now we allow people to set it to
'portable' or some random time.

Signed-off-by: William Stewart <zoidbergwill@gmail.com>