update_deps in Go
This removes one of two uses of the Python containerregistry

I wanted to use this in a repo of ours that uses rules_docker, and
didn't want anything that needs Python 2, so I thought moving this
to Go was the right approach.

Having a public binary for this would also be super useful for
encouraging folks to use digests instead of tags.