Allow setting compression method and options.
This makes it easier for users to specify fast compression to reduce
build times. One way to use it is to depend on compilation mode:

```
container_image(
    name = "my-image",
    compression_options = select({
        "@io_bazel_rules_docker//:optimized": [],
        "//conditions:default": ["--fast"],
    })
)
```

The only supported method is "gzip", but in future, zstd could be an
alternative compression method:
https://github.com/opencontainers/image-spec/blob/master/layer.md

zstd(1) also accepts the `--fast` flag: https://www.mankier.com/1/zstd

Fixes #1425.