Expose architecture in container_image
container_image builds all images labeled as amd64 despite the
possibility of it being intended for another architecture, without a way
to overwrite this value. Aside from being confusing to a user, this could
also be an issue when using said images as part of a manifest list, given
the wrong label won't pull the proper image to the system.

This patch exposes an architecture attribute that can be used by an user
in order to properly label the container image. The default is still amd64
in case it is not set.

It also adds a sh_test called architecture_test to verify that when a
labeled is set it was actually be properly modified in the config of the
target image.

Signed-off-by: Murilo Fossa Vicentini <muvic@linux.ibm.com>