Fix CI failures due to Clang 10 -> 11 upgrade
After Bazel container was updated to include Clang 11.0.0, this GCB test
started to fail because the toolchain config it uses is pinned to an
older Bazel container but the Bazel container the test runs in uses
"latest". From what I can tell, the test is meant to ensure RBE supports
the exec properties that enable sibling docker so I don't think we're
losing test coverage.