Add experimental_tarball_format to container_image
This adds an attribute that controls the format of the image tarball
produced when building container_image as "${name}.tar".

The current behavior is a legacy, uncompressed format that is compatible
with older versions of docker but suboptimal for most use cases. I have
retained the default behavior by setting the default value to "legacy",
but callers can opt into a newer format by setting this to "compressed".

In the future, we may want to add support for an OCI Image Layout as
well.