Revert "Invite maintainers from community (#1750)"
We got the new maintainers here now, and don't need to advertise this way.

We'll surely add more maintainers from the community over time, but can bootstrap that the normal way by promoting ppl who are active and helpful in the repo.

This reverts commit a4d580ee357b3e40e80becb311ea88214b70e0db
and moves the config file to .github/ folder to improve tidiness in the root