update_deps: Fix check for tag not found
This script finds reproducable hashes for pulling containers from
distroless. Since the last time this was run it seems that they
have stopped pushing a :debug tag for some images (static). This
makes sense as static is just an image with some standard openssl
and user file deps that do not have debug versions.

The code that implements update_deps is written to recover from
this situation (seemingly anticipating it) but it looks like at
some point the library we are using for interfacing with
container registries has changed. It now produces messages in
a slightly different format. I've updated the logic to be more
permissive of these failures. In the long run we should try to
land a PR in go-containerregistry to make this check easier.