Set local to true for dockerfile_build repo rule
Before this patch, any change to the Dockerfile won't trigger a
repository refetch. Some folks are using `bazel clean --expunge` to
workaround the issue.

This patch fixed the issue by setting the `local` field to be true for
the repository rule, which indicates that this rule fetches everything
from the local system and should be reevaluated at every fetch.