Vendor a GPG public key
The remote server we fetched it from was providing two different files (differing by a whitespace)
which caused the content hash to be unstable, making CI fail most of the time.

Note this hash was changed in the last month by
https://github.com/bazelbuild/rules_docker/commit/08cddcccfde27e014a9fbc01f53edac50e17e899
so the instability has been here for a while.