Update tarinfo's PAX header for extended/long file names.
tar can only support file names that are 99 characters long by
default. An extension to the command in POSIX.1-2001 allows an
extended set of headers to be specified which allows for
unlimited length file names. Python's tarfile module uses these
headers when writing new files into a tar. In this we set both
the TarInfo name and pax_header path when we've loaded a tar
file that contained them.