Add registry_file attribute to the container_pull rule
This change adds the attribute `registry_file` to the `container_push` rule. This allows all of the destination parameters (`registry, repository, tag`) to be supplied as either Label or String. The registry attribute pair follows the behaviour of `repository`, by requiring the String value to always be provided, even if overridden by the `repository_file`.

```python
container_push(
    name = "push_image",
    image = ":application_image",
    registry = "",
    registry_file = ":test.registry",
    repository = "docker/test",
    tag = "latest",
)
```

Passing the `registry` as a Label is intended for cases where the list of possible registries is not a fixed set. This is common in AWS (Amazon Web Services) where the registry for each account is based on the AWS Account ID and region. Publishing in this kind of infrastructure can mean that production/release artifacts are located within one "Artifacts" account, while developer sandbox artifacts can be spread across multiple sandbox accounts. Allowing for the `registry` to be a Label allows for the registry to be calculated based on the current working sandbox.