perf: remove no-remote-cache execution requirements (#2043)
These were introduced to reduce load on a remote-cache instance to avoid network saturation.
A month later, a feature was added in one remote-cache implementatation which provides a different fix:
https://github.com/buchgr/bazel-remote/issues/440 rejects large input files on upload.

In practice, while these action do often produce huge outputs, they are also slow to re-execute.
In many cases it's worth it to use a remote-cache for RunAndCommitLayer in particular to avoid a local rebuild
even though it's a large network fetch.
Currently users can't configure this because we've hardcoded the values. If they do want to keep the
no-remote-cache execution requirement, they can do this via a tag (provided they opt-in to
experimental_allow_tags_propagation, see https://github.com/bazelbuild/bazel/issues/8830)

https://github.com/bazelbuild/rules_docker/pull/1856#issuecomment-1060240919 is an example of a user
asking for these to be removed.