[layer_tools/generate_args_for_image] Inject extra argument to the given callback
When using a transition we need to give a callback function
to compute a path specific for this transition.
This information is not available from the context (ctx)
and the given file.
We need to have a way to inject this extra information.

Example with an architecture specific transition:

```starlark
def _to_arch_path(ctx, f, arch = None):
    """Returns the architecture specific path of the given file f."""
    return paths.join(
        paths.dirname(f.short_path),
        arch,
        paths.basename(f.short_path),
    )

def _xxx_impl(ctx):

    ...

    # image is an attribute with a 1:2+ transition
    for image_target in ctx.attr.image:
        image = ... [ImageInfo].container_parts
        arch = ...

        ...

        img_args, img_inputs = generate_args_for_image(
            ctx, image, _to_arch_path, arch = arch,
        )

        ...

        symlinks.update({
            _to_arch_path(ctx, i, arch = arch): i
            for i in img_inputs
        })

    ...

    runfiles = ctx.runfiles(
        files = ...
        symlinks = symlinks
    )
```