Set MaxIdleConnsPerHost for pushes (#2132)
This should reduce "use of closed network connection" errors due to too
many idle connections. Because we only expect to reach a couple hosts
during pushes anyway, this should be safe.

Currently, the value defaults to 2. This change sets it to half the
total IdleConnsPerHost, which defaults to 100.

New value is 100 / 2 = 50.