De-duplicate run args passed to bazel image targets.
When 'bazel run' is called, it expects that the args specified in the BUILD
file are not known to the run command and will prepend them to all other
arguments.

However, we have already baked those commands into the image, because users
expect that 'args' in the case of an image target will be included in the
image.  This templated script de-duplicates the args specified in the BUILD
file by passing itself the number of args already included in the image.