Ensure baseManifest is cleared in compat Reader.loadMetadata()
If a Reader is reused, as it is in join_layers, the execution of loadForeignLayers may
incorrectly fail for a tag with no base manifest due to the reader previously being used
on a tag with a base manifest. This failure is caused by attempting to use a base
manifest for the previous tag with the config for the current tag. Whether this would fail
would depend on the order in which the tags were loaded and the number of layers in each,
since loadForeignLayers only checked that the config has at least as many layers as the
base manifest. The fix is to ensure that baseManifest is cleared every time a new ReadImage
is executed.