chore: add maintenance status to top of README
14 out of 25 issues on the first page of https://github.com/bazelbuild/rules_docker/issues
have no replies.

Issues like #2213 show that releases are stalled right now.

https://github.com/bazelbuild/rules_docker/pulse/monthly shows four
commits touching four files in the last month, one of those was me
removing an external docs link.

I discussed this with the maintainers on the discussion thread, and we agreed to put up a message for now to better set user expectations.
I also alerted the Bazel team to this change in the SIG meeting on March 7, see
https://docs.google.com/document/d/1YGCYAGLzTfqSOgRFVsB8hDz-kEoTgTEKKp9Jd07TJ5c/edit#

To be clear: I greatly appreciate the effort that the current two
maintainers have made, this is not a vote-of-no-confidence in their
engineering abilities or desire to make this repo great. It's just
acknowledging the current state of the resources (time and funding)
available for this project at the present time.