Support runfiles root_symlinks in images
Fixes symlink support and adds root_symlink support. Previously I was
seeing that symlinks (regardless of whether the target file existed in
runfiles.file), did not point to the correct location in the image.

For that reason, we use `_final_file_path` instead of `layer_file_path`.
This seems to work, but I am not sure if I'm missing a use-case here.

The overall method we take here is to create real files where the
[root_]symlink is declared if the target `File` is not seen in
`file_map`.

This PR also adds some extra test functionality to check that symlinks
point at the expected location.